Replace ioutil.TempFile with custom function supporting permissions
The `renameio.WriteFile` function uses `(*os.File).Chmod` to apply the
file permissions given as a parameter. In doing so it doesn't respect
the umask.

To respect the umask it's necessary to specify the desired mode when
creating the file. To that end this change introduces a function
generating temporary files while also permitting the mode to be
controlled. Go's own `ioutil.TempFile` function always sets the file
mode to `0o600 & ^umask`, so usually 0o600.

`renameio.WriteFile` is not yet updated--a version of
`renameio.WriteFile` accepting permissions will be necessary for that.