Add file creation function receiving options (#35)
The `renameio.WriteFile` function uses `(*os.File).Chmod` to apply file
permissions without considering the umask. To control the permissions
used for new and replaced files more closely a new factory function is
added which receives option values.

Options allow the use of static permissions ignoring the umask (as
implemented by the existing `TempFile` function), permissions respecting
the umask (implemented by giving the mode to `os.OpenFile`) and for
attempting to look up the permissions of an existing file. The latter is
useful when overwriting, but by its nature it can't be atomic.

`renameio.WriteFile` is not yet updated--that'll follow in a separate
change including tests.