Respect umask in WriteFile
Go's own ioutil.WriteFile function passes the permissions directly to
the underlying system call which in turn respects the umask. The
previous implementation of "renameio.WriteFile" would always call
chmod(2), ignoring the umask. The recently added options for pending
files allow for the construction of a pending file while respecting the
umask.