Add IgnoreUmask option
There was already an option to ignore the umask when setting
permissions, "WithStaticPermissions". The "IgnoreUmask" option in
combination with "WithPermissions" is equivalent to using
"WithStaticPermissions".

For equivalence with Go's built-in "ioutil.WriteFile" function
renameio's "WriteFile" function should respect the umask. That's
a change in behaviour--currently the umask is ignored. To allow users to
continue to use the previous behaviour the newly added "IgnoreUmask"
option enables simpler usage where the permissions don't have to be
specified twice.

Plain usage respecting umask (eventually):

  renameio.WriteFile(…, 0o644)

Ignoring umask without "IgnoreUmask":

  renameio.WriteFile(…, 0o644, renameio.WithStaticPermissions(0o644))

Using "IgnoreUmask":

  renameio.WriteFile(…, 0o644, renameio.IgnoreUmask())